+++ HISTORY.md
@@ -1,5 +1,11 @@
 # Changelist history
 
+## V0.5.0
+- Added spam filter
+- Added db query for getting spam messages from db
+- Updated default message handler
+- Added new fields to config
+
 ## V0.4.2
 - Added HISTORY.md
 - Updated README.md